Minor update to drive slot filter

+++ app/data/format/drive-slot.filter.js
@@ -6,7 +6,7 @@
     .filter('driveSlot', ['$filter', function($filter) {
       return function (data) {
         if(typeof data === 'number' || typeof data === 'string') {
-          var number = typeof data === 'number' ? data : parseFloat(data),
+          var number = typeof data === 'number' ? data : parseInt(data),
             validNumber = number || number === 0;
 
           if(data === -1) { return 'internal'; }
